moved translation to menu
