+++ index.html
@@ -19,8 +19,8 @@
   <link
     href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&family=Rubik:ital,wght@0,300..900;1,300..900&display=swap"
     rel="stylesheet">
-  <link rel="icon" href="/public/img/icons/OK.png" type="image/png">
-  <link rel="shortcut icon" href="/public/img/icons/OK.png" type="image/png">
+  <link rel="icon" href="/img/icons/OK.png" type="image/png">
+  <link rel="shortcut icon" href="/img/icons/OK.png" type="image/png">
   <title>Nice Gadgets</title>
 </head>
