+++ src/components/Header/Header.tsx
@@ -6,24 +6,16 @@ import { Favorites } from '../Favorites';
 import { Cart } from '../Cart';
 import { Menu } from '../../modules/Menu';
 import { Settings } from '../Settings';
-import { useTranslation } from 'react-i18next';
 
 export const Header: React.FC = () => {
   const [isMenuOpen, setIsMenuOpen] = useState<boolean>(false);
   const [isSettingsOpen, setIsSettingsOpen] = useState<boolean>(false);
-  const { i18n } = useTranslation();
 
   const handleToggleMenu = () => {
     setIsMenuOpen(!isMenuOpen);
   };
   const toggleSettings = () => {
     setIsSettingsOpen(prevState => !prevState);
-  };
-
-  const changeLanguage = () => {
-    const newLanguage = i18n.language === 'en' ? 'ua' : 'en';
-    i18n.changeLanguage(newLanguage);
-    localStorage.setItem('language', newLanguage);
   };
 
   return (
@@ -39,11 +31,6 @@ export const Header: React.FC = () => {
       </div>
 
       <div className={classes.Header__right}>
-        <div className={classes.nav_LngWrap}>
-          <button onClick={changeLanguage} className={classes.nav_LngBtn}>
-            {i18n.language === 'en' ? 'UA' : 'EN'}
-          </button>
-        </div>
         <div className={classes.Header__settings}>
           <Settings isOpen={isSettingsOpen} onToggle={toggleSettings} />
         </div>
